Add sentence transformers and app for deduplication
