+++ deduplication/consts.py
@@ -1,0 +1,5 @@
+
+EMBEDDING_GLOVE_6B = 'average_word_embeddings_glove.6B.300d'
+EMBEDDING_GLOVE_840B = 'average_word_embeddings_glove.840B.300d'
+EMBEDDING_PARAPHRASE_MINI_LM_V2 = 'paraphrase-MiniLM-L3-v2'
+EMBEDDING_ALL_MPNET_V2 = 'all-mpnet-base-v2'
