+++ deduplication/tasks/indexing.py
@@ -1,0 +1,7 @@
+from sentence_transformers import SentenceTransformer
+
+from core.models import Lead, Project
+
+
+def create_lead_indices():
+    pass
