+++ pyproject.toml
@@ -16,6 +16,8 @@ pytest = "^7.2.0"
 coverage = "^6.5.0"
 psycopg2 = "^2.9.5"
 pandas = "^1.5.1"
+torch = "1.12.1"
+sentence-transformers = "2.2.2"
 
 [tool.poetry.dev-dependencies]
 mypy = "^0.990"
